+++ src/main/java/org/chzzk/howmeet/domain/regular/member/repository/MemberRepository.java
@@ -1,4 +1,4 @@
-package org.chzzk.howmeet.domain.regular.auth.repository;
+package org.chzzk.howmeet.domain.regular.member.repository;
 
 import org.chzzk.howmeet.domain.regular.auth.entity.Member;
 import org.springframework.data.jpa.repository.JpaRepository;
